refactor(dwn-sdk-js): decompose RecordsWrite into focused modules
Split signing/crypto and query/fetch helpers out of records-write.ts:

- records-write-signing.ts (~143 lines): createEncryptionProperty,
  createAttestation, createSignerSignature, validateAttestationIntegrity
- records-write-query.ts (~139 lines): getInitialWrite,
  verifyEqualityOfImmutableProperties, getAttesters, fetchNewestRecordsWrite,
  fetchInitialRecordsWrite, fetchInitialRecordsWriteMessage

The main records-write.ts retains the class with constructor, factory methods
(parse, create, createFrom), instance methods (sign, signAsOwner, signAsOwnerDelegate,
encryptSymmetricEncryptionKey, validateIntegrity, constructIndexes, authorize*),
getters, and getEntryId/isInitialWrite. Static delegate methods preserve
backward compatibility for all public API symbols.

Uses lazy dynamic import in records-write-query.ts to break the circular
dependency (RecordsWrite -> records-write-query -> RecordsWrite).
